requirements.txt
asgiref==3.8.1
Django==4.2
django-extensions==3.2.3
django-filter==24.3
django-htmx==1.18.0
django-widget-tweaks==1.5.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
